+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add new `TextLinkDetector` for parsing and detecting links given a string [#2984](https://github.com/GetStream/stream-chat-swift/pull/2984)
 ### 🐞 Fixed
 - Fix message link preview showing empty space when no metadata [#2984](https://github.com/GetStream/stream-chat-swift/pull/2984)
+- Fix threading issue in `ConnectionRepository` [#2985](https://github.com/GetStream/stream-chat-swift/pull/2985)
 
 ## StreamChatUI
 ### ✅ Added